Add PermissionsMixin to User model.
