+++ physionet-django/project/modelcomponents/activeproject.py
@@ -141,6 +141,9 @@ class ActiveProject(Metadata, UnpublishedProject, SubmissionInfo):
         50: 'Awaiting authors to approve publication.',
         60: 'Awaiting editor to publish.',
     }
+    
+    class Meta:
+        default_permissions = ()
 
     def storage_used(self):
         """
