+++ physionet-django/project/modelcomponents/archivedproject.py
@@ -22,5 +22,8 @@ class ArchivedProject(Metadata, UnpublishedProject, SubmissionInfo):
     # Where all the archived project files are kept
     FILE_ROOT = os.path.join(settings.MEDIA_ROOT, 'archived-projects')
 
+    class Meta:
+        default_permissions = ()
+
     def __str__(self):
         return ('{0} v{1}'.format(self.title, self.version))
