+++ physionet-django/project/modelcomponents/legacy.py
@@ -34,6 +34,9 @@ class LegacyProject(models.Model):
     contact_affiliations = models.CharField(max_length=150, default='MIT')
     contact_email = models.EmailField(max_length=255, default='webmaster@physionet.org')
 
+    class Meta:
+        default_permissions = ()
+
     # Put the references as part of the full description
 
     def __str__(self):
