+++ physionet-django/user/models.py
@@ -6,10 +6,10 @@ from datetime import timedelta
 from django.conf import settings
 from django.contrib import messages
 # from django.contrib.auth. import user_logged_in
-from django.contrib.auth import get_user_model, signals
-from django.contrib.auth.models import AbstractBaseUser, BaseUserManager
+from django.contrib.auth import signals
+from django.contrib.auth.models import AbstractBaseUser, BaseUserManager, PermissionsMixin
 from django.core.exceptions import ObjectDoesNotExist, ValidationError
-from django.core.validators import EmailValidator, FileExtensionValidator, integer_validator, validate_integer
+from django.core.validators import EmailValidator, FileExtensionValidator
 from django.db import DatabaseError, models, transaction
 from django.db.models import CharField
 from django.db.models.functions import Lower
@@ -297,8 +297,7 @@ class UserManager(BaseUserManager):
         user.set_password(password)
         user.save(using=self._db)
 
-        profile = Profile.objects.create(user=user, first_names=first_names,
-                                         last_name=last_name)
+        Profile.objects.create(user=user, first_names=first_names, last_name=last_name)
         return user
 
     def create_superuser(self, email, password, username):
@@ -320,7 +319,7 @@ def validate_unique_email(email):
             code='email_not_unique',)
 
 
-class User(AbstractBaseUser):
+class User(AbstractBaseUser, PermissionsMixin):
     """
     The user authentication model
     """
@@ -349,9 +348,11 @@ class User(AbstractBaseUser):
     RELATIVE_FILE_ROOT = 'users'
     FILE_ROOT = os.path.join(UserFiles().file_root, RELATIVE_FILE_ROOT)
 
-    def is_superuser(self):
-        return (self.is_admin,)
+    objects = UserManager()
 
+    class Meta:
+        default_permissions = ('view',)
+    
     # Mandatory methods for default authentication backend
     def get_full_name(self):
         return self.profile.get_full_name()
@@ -362,16 +363,6 @@ class User(AbstractBaseUser):
     def __str__(self):
         return self.username
 
-    objects = UserManager()
-
-    # Mandatory attributes for using the admin panel
-    def has_perm(self, perm, obj=None):
-        "Does the user have a specific permission?"
-        return True
-
-    def has_module_perms(self, app_label):
-        "Does the user have permissions to view the app `app_label`?"
-        return True
 
     @property
     def is_staff(self):
@@ -421,6 +412,10 @@ class UserLogin(models.Model):
     login_date = models.DateTimeField(auto_now_add=True, null=True)
     ip = models.CharField(max_length=50,  blank=True, default='', null=True)
 
+    class Meta:
+        default_permissions = ()
+
+
 def update_user_login(sender, **kwargs):
     user = kwargs.pop('user', None)
     request = kwargs.pop('request', None)
@@ -458,6 +453,9 @@ class AssociatedEmail(models.Model):
     # Time limit for verification: maximum number of days after
     # 'added_date' during which 'verification_token' may be used.
     VERIFICATION_TIMEOUT_DAYS = 7
+
+    class Meta:
+        default_permissions = ()
 
     def __str__(self):
         return self.email
@@ -548,6 +546,9 @@ class LegacyCredential(models.Model):
 
     revoked_datetime = models.DateTimeField(null=True)
 
+    class Meta:
+        default_permissions = ()
+
     def __str__(self):
         return self.email
 
@@ -598,6 +599,9 @@ class Profile(models.Model):
 
     MAX_PHOTO_SIZE = 2 * 1024 ** 2
 
+    class Meta:
+        default_permissions = ()
+
     def __str__(self):
         return self.get_full_name()
 
@@ -640,33 +644,13 @@ class Orcid(models.Model):
     token_scope = models.CharField(max_length=50, default='', blank=True)
     token_expiration = models.DecimalField(max_digits=50, decimal_places=40, default=0)
 
+    class Meta:
+        default_permissions = ()
+
     @staticmethod
     def get_orcid_url():
         return settings.ORCID_DOMAIN
 
-class DualAuthModelBackend():
-    """
-    This is a ModelBacked that allows authentication with either a username or an email address.
-
-    """
-    def authenticate(self, request, username=None, password=None):
-        if '@' in username:
-            kwargs = {'email': username.lower()}
-        else:
-            kwargs = {'username': username.lower()}
-        try:
-            user = get_user_model().objects.get(**kwargs)
-            if user.check_password(password):
-                return user
-        except User.DoesNotExist:
-            logger.error('Unsuccessful authentication {0}'.format(username.lower()))
-            return None
-
-    def get_user(self, user_id):
-        try:
-            return get_user_model().objects.get(pk=user_id)
-        except get_user_model().DoesNotExist:
-            return None
 
 class CredentialApplication(models.Model):
     """
@@ -781,6 +765,9 @@ class CredentialApplication(models.Model):
     responder_comments = models.CharField(max_length=500, default='',
         blank=True)
     revoked_datetime = models.DateTimeField(null=True)
+
+    class Meta:
+        default_permissions = ('change',)
 
     def file_root(self):
         """Location for storing files associated with the application"""
@@ -1003,6 +990,9 @@ class CredentialReview(models.Model):
     responder_comments = models.CharField(max_length=500, default='',
                                           blank=True)
 
+    class Meta:
+        default_permissions = ()
+
 
 class CloudInformation(models.Model):
     """
@@ -1013,3 +1003,6 @@ class CloudInformation(models.Model):
     gcp_email = models.OneToOneField('user.AssociatedEmail', related_name='gcp_email',
         on_delete=models.SET_NULL, null=True)
     aws_id = models.CharField(max_length=60, null=True,  blank=True, default=None)
+
+    class Meta:
+        default_permissions = ()
